Rename Absolute test file to Instant
This was forgotten in the rename.
